+++ packages/adherev-util-communication-ajax/src/ajax.ts
@@ -63,16 +63,7 @@ function createXHR() {
  * getDefaultConfig - 返回构造函数config的默认值
  * @return IConfig
  */
-function getDefaultConfig(): IConfig & {
-  loading: {
-    // 是否显示遮罩
-    show: boolean;
-    // 遮罩的内容
-    text: string;
-    // 遮罩的元素
-    el: HTMLElement;
-  };
-} {
+function getDefaultConfig(): IConfig {
   return {
     timeout: Ajax.TIMEOUT,
     withCredentials: true,
@@ -108,6 +99,7 @@ function getDefaultConfig(): IConfig & {
           break;
       }
     },
+    mock: false,
     // loading的配置
     loading: {
       // 是否显示遮罩
@@ -117,6 +109,12 @@ function getDefaultConfig(): IConfig & {
       // 遮罩的元素
       el: document.body,
     },
+    onBeforeResponse: () => {},
+    dataKey: 'data',
+    messageKey: 'message',
+    codeKey: 'code',
+    codeSuccess: 200,
+    showWarn: true,
   };
 }
 
@@ -211,8 +209,8 @@ function onreadystatechange({
         // 只有application/json才进行三大值的判断
         const jsonObj = JSON.parse(xhr.responseText);
 
-        if (showWarn && jsonObj[codeKey] !== codeSuccess) {
-          warnInfo(intl.tv('提示'), jsonObj[messageKey]);
+        if (showWarn && codeKey in jsonObj && jsonObj[codeKey] !== codeSuccess) {
+          warnInfo(intl.v('提示'), jsonObj[messageKey]);
         }
 
         resolve(resolveData({ show, data: jsonObj, indicator }));
@@ -293,7 +291,7 @@ function sendPrepare(
 
   const defaultLoadingText = `${intl.tv('加载中')}...`;
 
-  const { show = false, text = defaultLoadingText, el = document.body } = loading;
+  const { show = false, text = defaultLoadingText, el = document.body } = loading!;
 
   // 显示loading
   if (show) {
@@ -363,14 +361,13 @@ function sendPrepare(
     if (!Util.isEmpty?.(data) && Util.isRef?.(data) && method !== ('get' || 'GET')) {
       if (
         !(
-          // @ts-ignore
-          (
-            'form' in data &&
-            'data' in data &&
-            !Util.isEmpty?.(data.form) &&
-            !Util.isEmpty?.(data.data) &&
-            data.form instanceof HTMLFormElement
-          )
+          'form' in
+            // @ts-ignore
+            data &&
+          'data' in data &&
+          !Util.isEmpty?.(data.form) &&
+          !Util.isEmpty?.(data.data) &&
+          data.form instanceof HTMLFormElement
         )
       ) {
         // console.log('默认设置Content-Type', `${Ajax.CONTENT_TYPE_APPLICATION_JSON};charset=utf-8`);
@@ -439,25 +436,31 @@ function getSendParams({ data, contentType }) {
 
   // console.log('getSendParams', data, contentType);
 
-  // application/json
-  if (contentType.indexOf(Ajax.CONTENT_TYPE_APPLICATION_JSON) === 0 && Util.isRef?.(data)) {
+  /**
+   * application/json
+   */
+  if (contentType.startsWith(Ajax.CONTENT_TYPE_APPLICATION_JSON) && Util?.isRef?.(data)) {
     // console.log('数据需要被转换成JSON字符串', JSON.stringify(data));
     return JSON.stringify(data);
   }
 
-  // application/x-www-form-urlencoded
+  /**
+   * application/x-www-form-urlencoded
+   */
   if (
-    contentType.indexOf(Ajax.CONTENT_TYPE_APPLICATION_X_WWW_FORM_URLENCODED) === 0 &&
-    Util.isObject?.(data)
+    contentType.startsWith(Ajax.CONTENT_TYPE_APPLICATION_X_WWW_FORM_URLENCODED) &&
+    Util?.isObject?.(data)
   ) {
     // console.log('application/x-www-form-urlencoded转换', JSON.stringify(data));
     return Array.from(Object.keys(data))
-      .map(k => encodeURIComponent(`${k}=${data[k]}`))
+      .map((k) => encodeURIComponent(`${k}=${data[k]}`))
       .join('&');
   }
 
-  // multipart/form-data
-  if (contentType.indexOf(Ajax.CONTENT_TYPE_MULTIPART_FORM_DATA) === 0 && Util.isObject?.(data)) {
+  /**
+   * multipart/form-data
+   */
+  if (contentType.startsWith(Ajax.CONTENT_TYPE_MULTIPART_FORM_DATA) && Util?.isObject?.(data)) {
     // console.log('multipart/form-data转换');
     // console.log('form', data.form);
 
@@ -470,6 +473,16 @@ function getSendParams({ data, contentType }) {
 
     return formData;
   }
+
+  /**
+   * text/plain
+   */
+  if (contentType.startsWith(Ajax.CONTENT_TYPE_TEXT_PLAIN)) {
+    if (Util?.isString?.(data)) return data;
+    if (Util?.isObject?.(data)) return JSON.stringify(data);
+  }
+
+  return data.toString();
 }
 
 /**