+++ packages/adherev-ui-conditionalrender/changelog/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.4
+
+***
+
+2022-04-05
+
+* 加入conditionalRender和conditionalArr两个方法
+
 # 1.0.8
 
 ***